Enables GPU for Truncated Newton code.

+++ src/precision.h
@@ -11,6 +11,6 @@
 #define PRECISION_H
 
 typedef double real;
-typedef unsigned long int integer;
+typedef unsigned int integer;
 
 #endif
